[ᚬmaster] Rc/v0.24.8

+++ example/src/main/java/org/nervos/ckb/transaction/CellCollector.java
@@ -32,8 +32,8 @@ public class CellCollector {
       throws IOException {
     List<String> cellOutputsData = new ArrayList<>();
     for (int i = 0; i < cellOutputs.size() - 1; i++) {
-      int size = cellOutputs.get(i).occupiedCapacity("0x");
-      if (BigInteger.valueOf(size).compareTo(Numeric.toBigInt(cellOutputs.get(i).capacity)) > 0) {
+      BigInteger size = cellOutputs.get(i).occupiedCapacity("0x");
+      if (size.compareTo(Numeric.toBigInt(cellOutputs.get(i).capacity)) > 0) {
         throw new IOException("Cell output byte size must not be bigger than capacity");
       }
       cellOutputsData.add("0x");
